update test format and update entrypoint tests
